+++ tests/testthat/test-GGE_Model.R
@@ -6,7 +6,7 @@ library(agridat)
 data(yan.winterwheat)
 
 
-GGE1 <- GGEmodel(yan.winterwheat, genotype="gen",environment="env", response="yield", rep=NULL, centering = "tester")
+GGE1 <- GGEmodel(yan.winterwheat, genotype="gen",environment="env", response="yield", rep=NULL)
 
 dat2 <- yan.winterwheat
 dat2[1,3]<-NA
@@ -16,17 +16,17 @@ dat2[2,3]<-NA
 
 test_that("Several tests for GGE_Model function", {
 
-  expect_that(GGEmodel(yan.winterwheat,genotype="gen",environment="env", response="yield", centering = "tester"), is_a("GGEModel") )
+  expect_that(GGEmodel(yan.winterwheat,genotype="gen",environment="env", response="yield"), is_a("GGEModel") )
 
-  expect_error(GGEmodel(genotype="gen",environment="env", response="yield", centering = "tester"),
+  expect_error(GGEmodel(genotype="gen",environment="env", response="yield"),
                "Need to provide Data data frame")
 
-  expect_equal(GGEmodel(yan.winterwheat,genotype="gen",environment="env", response="yield", centering = "tester"),
+  expect_equal(GGEmodel(yan.winterwheat,genotype="gen", environment="env", response="yield"),
               GGEmodel(yan.winterwheat))
 
   expect_identical(yan.winterwheat, na.omit(yan.winterwheat))
 
-  expect_error(GGEmodel(dat2,genotype="gen",environment="env", response="yield", centering = "tester"),
+  expect_error(GGEmodel(dat2,genotype="gen",environment="env", response="yield"),
                "Missing data in input data frame, run the imputation function first to complete the data set")
 
   # data types correct